implement end method

+++ src/common.rs
@@ -1,10 +1,10 @@
 pub struct Settings {
     pub period: u64,
     pub current_avg_points: f64,
-    pub current_points_degree: u128,
+    pub current_points_degree: f64,
     pub current_contribution_value: f64,
     pub eth: u128,
-    pub tokens: u128,
+    pub tokens: f64,
 }
 
 
@@ -15,10 +15,10 @@ impl Settings {
         Settings {
             period: 1,
             current_avg_points: 1.5,
-            current_points_degree: 10,
+            current_points_degree: 10.0,
             current_contribution_value: 10.0,
             eth: 0,
-            tokens: 0,
+            tokens: 0.0,
         }
     }
